% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crimes.R
\name{importData}
\alias{importData}
\title{Import data from Istat crimes dataset}
\usage{
importData()
}
\value{
data.frame
}
\description{
This function returns the italian crimes from 2014 to 2018.
Data source: http://dati.istat.it/Index.aspx?DataSetCode=DCCV_DELITTIPS.
}
